+++ xfv/src/cohesive_calculation/linearpercent.py
@@ -24,9 +24,9 @@ class LinearPercent(CohesiveCalculationModel) :  # pylint: disable=too-few-publi
         :cohesive_model: type of cohesive model 
         """
 
-        critical_strength = abs(stress[ind, 0])
-        dissipated_energy = energy[ind]*mass[ind]*cohesive_model.purcentage
-        critical_separation = 2.*dissipated_energy/(critical_strength*section)
+        critical_strength = float(abs(stress[ind, 0]))
+        dissipated_energy = float(energy[ind]*mass[ind]*cohesive_model.purcentage)
+        critical_separation = float(2.*dissipated_energy/(critical_strength*section))
         print('critical strength =', critical_strength)
         print('critical separation =', critical_separation)
         print('cohesive dissipated energy =', dissipated_energy)
